Adding visualization and the ability to customize the size of the parking lot
Added a configuration file and a class for working with it. The method of obtaining data from the configuration file when generating a parking lot has been prescribed. Added methods for visualizing the list of parked vehicles and a diagram of its location. Transport objects are passed to parking lot objects.

TODO:
- create a timer for vehicle exit from the parking lot
- create a vehicle exit method for the parking attendant
- describe the architecture and modules used
- describe the logic of the program